field types and row value types to separate files
