+++ baserowapi/__init__.py
@@ -1,5 +1,3 @@
 from .baserow import Baserow
-from .models import Table
-from .models import Row
-from .models import Filter
-from .models import Field, FieldList
+from .models import *
+from .validators.filter_validator import FilterValidator
