+++ baserowapi/models/__init__.py
@@ -1,4 +1,7 @@
-from .table import Table
-from .row import Row
 from .filter import Filter
-from .field import Field, FieldList
+from .row import Row
+from .table import Table
+
+# Import the submodules
+from .fields import *
+from .row_values import *
